requirements.txt
SQLAlchemy-Utils==0.32.9
buildbot-worker==0.9.3
buildbot-www==0.9.3
buildbot[tls]==0.9.3
mysql-python==1.2.5
netifaces==0.10.5
paramiko==2.0.1
pyaml==15.8.2
python-novaclient==2.35
raven==5.23.0
requests[security]==2.10.0
retrying==1.3.3
service_identity==16.0.0
txrequests==0.9.5 # required by the HTTPStep
xunitparser==1.3.3
twisted<17
